Add very ugly `continuous_single_pass()`
This is a drop-in replacement for the function
`MakeContinuousSpectrumSingPass` in the Mathematica module.

The current code is very very ugly and sub-optimal - it's a near direct
translation of the Mathematica code, with no thought given to code reuse
or legibility (I just wanted to make it work, first).

This should be changed into a more legible version, and I should give
some thought as to how the data should be organised at every stage.