Add data extraction functions
Mainly this adds the functions `independents()` and `probabilities()` to
the code, which should essentially give this all the functionality
contained in the Mathematica module this is based on.

The basic data structure used throughout for the data is a structured
`numpy.array`, which makes the manipulation of the different fields of
data significantly easier for the user.

Since the type of scan is encoded in the metadata, this can be
automatically detected, and there is no need for all the different
functions in the Mathematica file - the user only really needs `load()`
to read the data file, `independents()` to get the values of the
independent variables and `probabilities()` to get the excitation
probabilities and the associated errors.  This should make code re-use
much easier, and hopefully easier for the end user.

Some types of scan files may not yet be supported - I haven't looked
through the C# to see the different types of files that can be created.
Certainly windowed and continuous frequency scans are supported, as are
Rabi-type oscillation scans.

This commit is quite large - it can more accurately be seen as an
initial comit creating the port.  The previous commits in the git
history were more trying things out.