Allow load() to override number of shots per point
There was a problem in some files reporting 100 shots per point, when
they were actually taken at 400 shots per point.  This commit teachs
load() how to override the number of shots, subject to this override
dividing cleanly into the total number of data acquisitions in the file.

Since splitting up of the acquisitions into points doesn't happen until
probabilities() is called, this ought to be safe.