Add Axial model
At first this just adds the `fit.Axial` class, which can model the
probability of excitation of an ion in a thermal state as a function of
the detuning of the 729 laser from the carrier transition.

This can be used to estimate the probability of the ion being prepared
in the ground state of motion after sideband cooling.