Add XML pulse sequence designer
This can be used as a much more general alternative to the C#'s pulse
sequence designer, which now gives you the full power of Python to
create your sequences.

NOTE: This writes out XML, not the .hex files.  When I started writing
this, I didn't realise that the .hex files were the actual important
bit.  To use one of these XML files, you have to load up the XML file,
and then save it over itself using the pulse sequence designer to get
the C# to generate the hex.  In the future I will hope to add my own hex
generator so that this step can be bypassed.

For convenience, there is a high-level interface, which is still more
powerful than the C#, but has some limitations because it's designed to
make it easy to make the pulse sequences we currently make.

There is also a lower-level interface, which is essentially equivalent
to manually inputting each pulse in the C#.  Since this can be done
programmatically, you have full control over what you create, and it
should be significantly faster than using the GUI to do it (since you
can use Python loops and control flow to make the pulses).