Make cooling lasers active during commands
This oughtn't to be strictly necessary, but this matches what the C#
does.  The C# also has the low bit of the DDS profile set, which means
that it's interpreted as profile 6.  I didn't match this - there seems
to be no reason why I should, and there seems to be no reason why any of
the DDS pins should be set at all (which corresponds to profile 7 - the
DDS profile is the ones' complement of the three bits).