Tidy up pulse creation
The initial pass at FPGA hex file creation was very fast and loose - it
was largely just copied and pasted from the XML creation, so there was a
lot of duplication.  This simplifies the process, and reuses code that
should be reused.

In the future, I might add some addition parts of the pulse sequence
creator - I want to make it easier to tag what constitutes one "shot" of
the experiment, so it becomes easier to interleave multiple experiments
into the same hex file.

I might also add some methods for automatically inserting `send_data`
comamnds - there's no real need for the user to have to do that.  They
could just specify a strategy for doing it: pointwise; necessary;
etc; and let the program fill that in automatically.