Add communication with the DDS Arduino
This can now do all the same communication with the DDS Arduino that the
C# can do.  The only hardware that is missing from the Python now is the
FPGA communication (though that is by far the trickiest).  Once that is
in, it will (in theory) be possible to run experiments from this Python
code.  This was only meant to be data analysis code when I started the
package!

There is now a dependency on the `pyserial` package, which is
installable through the default channels on `conda`.