Support loading multiple files of the same type
A more perfect implementation would infer the python type and allow
different data file types to be loaded at once. This is not that
implementation.

I didn't import 'load' into the package namespace so as not to conflict
with the 'load' used for spectrum data files that's already there.
Ideally we would unify loading all file types under a single API in
the future.